Remove all forwardRef(), Merge Settings and PeriodSettings
