+++ packages/api/src/api-key/api-key.module.ts
@@ -1,6 +1,6 @@
 import { ConstantsProvider } from '../constants/constants.provider';
 import { EventLogModule } from '../event-log/event-log.module';
-import { Module, forwardRef } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ApiKeyController } from './api-key.controller';
 import { ApiKeyService } from './api-key.service';
@@ -10,8 +10,7 @@ import { AuthModule } from '../auth/auth.module';
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: ApiKey.name, schema: ApiKeySchema }]),
-    forwardRef(() => EventLogModule),
-    forwardRef(() => AuthModule),
+    EventLogModule,
   ],
   controllers: [ApiKeyController],
   providers: [ApiKeyService, ConstantsProvider],
