+++ packages/api/src/auth/eth-signature.service.ts
@@ -1,16 +1,16 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { User } from '../users/schemas/users.schema';
 import { JwtService } from '@nestjs/jwt';
-import { UsersService } from '../users/users.service';
 import { JwtPayload } from './interfaces/jwt-payload.interface';
 import { LoginResponseDto } from './dto/login-response.dto';
 import { EventLogService } from '../event-log/event-log.service';
 import { EventLogTypeKey } from '../event-log/enums/event-log-type-key';
 import { ApiException } from '../shared/exceptions/api-exception';
-import { randomBytes } from 'crypto';
 import { errorMessages } from '../shared/exceptions/error-messages';
 import { HOSTNAME_TEST } from '../constants/constants.provider';
 import { ethers } from 'ethers';
+import { User, UserDocument } from '../users/schemas/users.schema';
+import { Model } from 'mongoose';
+import { InjectModel } from '@nestjs/mongoose';
 
 @Injectable()
 /**
@@ -18,37 +18,11 @@ import { ethers } from 'ethers';
  */
 export class EthSignatureService {
   constructor(
-    private readonly usersService: UsersService,
     private readonly jwtService: JwtService,
     private readonly eventLogService: EventLogService,
+    @InjectModel(User.name)
+    private userModel: Model<UserDocument>,
   ) {}
-
-  /**
-   * Generates a nonce for the user and returns it.
-   *
-   * @param identityEthAddress Ethereum address of the user
-   * @returns User with updated nonce
-   */
-  async generateUserNonce(identityEthAddress: string): Promise<User> {
-    // Generate random nonce used for auth request
-    const nonce = randomBytes(10).toString('hex');
-
-    try {
-      // Find user by their Ethereum address, update nonce
-      const user = await this.usersService.findOneByEth(identityEthAddress);
-      return this.usersService.update(user._id, { nonce });
-    } catch (e) {
-      // No user found, create a new user
-      return this.usersService.create({
-        identityEthAddress,
-        rewardsEthAddress: identityEthAddress,
-        username: await this.usersService.generateValidUsername(
-          identityEthAddress,
-        ),
-        nonce,
-      });
-    }
-  }
 
   /**
    * Generates a login message that will be signed by the frontend user, and validated by the API.
@@ -76,9 +50,9 @@ export class EthSignatureService {
     signature: string,
     hostname: string,
   ): Promise<LoginResponseDto> {
-    let user;
+    let user: User;
     try {
-      user = await this.usersService.findOneByEth(identityEthAddress);
+      user = await this.userModel.findOne({ identityEthAddress }).lean();
     } catch (e) {
       // Throw UnauthorizedException instead of BadRequestException since
       // the user is not authenticated yet Nest.js defaults to that on
