+++ packages/api/src/database/interfaces/migration-context.interface.ts
@@ -1,5 +1,4 @@
 import { PeriodsService } from '../../periods/services/periods.service';
-import { PeriodSettingsService } from '../../periodsettings/periodsettings.service';
 import { PraiseService } from '../../praise/services/praise.service';
 import { QuantificationsService } from '../../quantifications/services/quantifications.service';
 import { SettingsService } from '../../settings/settings.service';
@@ -11,7 +10,6 @@ export interface MigrationsContext {
     usersService: UsersService;
     periodService: PeriodsService;
     settingsService: SettingsService;
-    periodSettingsService: PeriodSettingsService;
     quantificationsService: QuantificationsService;
   };
 }
