+++ packages/api/src/database/modules/app-migrations.module.ts
@@ -4,7 +4,6 @@ import { UserAccountsModule } from '../../useraccounts/useraccounts.module';
 import { UsersModule } from '../../users/users.module';
 import { EventLogModule } from '../../event-log/event-log.module';
 import { SettingsModule } from '../../settings/settings.module';
-import { PeriodSettingsModule } from '../../periodsettings/periodsettings.module';
 import { PraiseModule } from '../../praise/praise.module';
 import { QuantificationsModule } from '../../quantifications/quantifications.module';
 import { RequestContextModule } from 'nestjs-request-context';
@@ -25,7 +24,6 @@ export class AppMigrationsModule {
         CommunityModule,
         EventLogModule,
         PeriodsModule,
-        PeriodSettingsModule,
         PraiseModule,
         QuantificationsModule,
         RequestContextModule,
