+++ packages/api/src/quantifications/quantifications.module.ts
@@ -9,11 +9,9 @@ import { UserAccountsModule } from '../useraccounts/useraccounts.module';
 import { UsersModule } from '../users/users.module';
 import { PraiseModule } from '../praise/praise.module';
 import { PeriodsModule } from '../periods/periods.module';
-import { Module, forwardRef } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { QuantificationsController } from './quantitifcations.controller';
 import { QuantificationsExportService } from './services/quantifications-export.service';
-import { AuthModule } from '../auth/auth.module';
-import { ApiKeyModule } from '../api-key/api-key.module';
 import { ConstantsProvider } from '../constants/constants.provider';
 
 @Module({
@@ -21,13 +19,11 @@ import { ConstantsProvider } from '../constants/constants.provider';
     MongooseModule.forFeature([
       { name: Quantification.name, schema: QuantificationsSchema },
     ]),
-    forwardRef(() => PraiseModule),
-    forwardRef(() => SettingsModule),
-    forwardRef(() => PeriodsModule),
-    forwardRef(() => UsersModule),
+    PraiseModule,
+    SettingsModule,
+    PeriodsModule,
+    UsersModule,
     UserAccountsModule,
-    forwardRef(() => AuthModule),
-    ApiKeyModule,
   ],
   controllers: [QuantificationsController],
   providers: [
