+++ packages/api/src/settings/schemas/periodsettings.schema.ts
@@ -4,7 +4,7 @@ import { ExposeId } from '../../shared/decorators/expose-id.decorator';
 import { ApiProperty, ApiResponseProperty } from '@nestjs/swagger';
 import { Period } from '../../periods/schemas/periods.schema';
 import { Type } from 'class-transformer';
-import { Setting } from '../../settings/schemas/settings.schema';
+import { Setting } from './settings.schema';
 import { IsString } from 'class-validator';
 
 export type PeriodSettingDocument = PeriodSetting & Document;
