+++ packages/api/src/settings/settings.module.ts
@@ -1,22 +1,22 @@
 import { FileUtilsProvider } from '../settings/providers/file-utils.provider';
-import { Module, forwardRef } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ConstantsProvider } from '../constants/constants.provider';
 import { Setting, SettingSchema } from './schemas/settings.schema';
 import { SettingsController } from './settings.controller';
 import { SettingsService } from './settings.service';
 import { EventLogModule } from '../event-log/event-log.module';
-import { PeriodSettingsModule } from '../periodsettings/periodsettings.module';
-import { AuthModule } from '../auth/auth.module';
-import { ApiKeyModule } from '../api-key/api-key.module';
+import { Period, PeriodSchema } from '../periods/schemas/periods.schema';
+import { PeriodSetting } from './schemas/periodsettings.schema';
 
 @Module({
   imports: [
-    MongooseModule.forFeature([{ name: Setting.name, schema: SettingSchema }]),
-    forwardRef(() => EventLogModule),
-    forwardRef(() => PeriodSettingsModule),
-    forwardRef(() => AuthModule),
-    ApiKeyModule,
+    MongooseModule.forFeature([
+      { name: Setting.name, schema: SettingSchema },
+      { name: PeriodSetting.name, schema: PeriodSetting },
+      { name: Period.name, schema: PeriodSchema },
+    ]),
+    EventLogModule,
   ],
   controllers: [SettingsController],
   providers: [SettingsService, FileUtilsProvider, ConstantsProvider],
