+++ packages/api/src/settings/settings.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable, forwardRef } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
 import { Setting } from './schemas/settings.schema';
@@ -8,18 +8,23 @@ import { ApiException } from '../shared/exceptions/api-exception';
 import { EventLogService } from '../event-log/event-log.service';
 import { EventLogTypeKey } from '../event-log/enums/event-log-type-key';
 import { SettingGroup } from './enums/setting-group.enum';
-import { PeriodSettingsService } from '../periodsettings/periodsettings.service';
 import { validateSetting } from './utils/validate-setting';
 import { SettingsFilterDto } from './dto/settings-filter.dto';
 import { errorMessages } from '../shared/exceptions/error-messages';
+import { PeriodSetting } from './schemas/periodsettings.schema';
+import { SetPeriodSettingDto } from './dto/set-periodsetting.dto';
+import { Period } from '../periods/schemas/periods.schema';
+import { PeriodStatusType } from '../periods/enums/status-type.enum';
 
 @Injectable()
 export class SettingsService {
   constructor(
     @InjectModel(Setting.name)
     private settingsModel: Model<Setting>,
-    @Inject(forwardRef(() => PeriodSettingsService))
-    private periodSettingsService: PeriodSettingsService,
+    @InjectModel(PeriodSetting.name)
+    private periodSettingsModel: Model<PeriodSetting>,
+    @InjectModel(Period.name)
+    private periodModel: Model<Period>,
     private utils: FileUtilsProvider,
     private eventLogService: EventLogService,
   ) {}
@@ -28,8 +33,16 @@ export class SettingsService {
    * Convenience method to get the Settings Model
    * @returns
    */
-  getModel(): Model<Setting> {
+  getSettingsModel(): Model<Setting> {
     return this.settingsModel;
+  }
+
+  /**
+   * Convenience method to get the PeriodSettings Model
+   * @returns
+   */
+  getPeriodSettingsModel(): Model<PeriodSetting> {
+    return this.periodSettingsModel;
   }
 
   /**
@@ -190,11 +203,12 @@ export class SettingsService {
         .lean();
 
       if (generalSetting) {
-        setting =
-          await this.periodSettingsService.findOneBySettingIdAndPeriodId(
-            generalSetting._id,
-            periodId,
-          );
+        setting = await this.periodSettingsModel
+          .findOne({ period: periodId, setting: generalSetting._id })
+          .lean()
+          .populate('period')
+          .populate('setting')
+          .exec();
 
         if (!setting) {
           throw new ApiException(errorMessages.SETTING_NOT_FOUND);
@@ -217,5 +231,111 @@ export class SettingsService {
         group,
       })
       .lean();
+  }
+
+  async findAllPeriodSettings(
+    periodId: Types.ObjectId,
+  ): Promise<PeriodSetting[]> {
+    const settings = await this.periodSettingsModel
+      .find({ period: periodId })
+      .lean()
+      .populate('period')
+      .populate('setting')
+      .exec();
+    return settings.map((setting) => new PeriodSetting(setting));
+  }
+
+  async findOneBySettingIdAndPeriodId(
+    settingId: Types.ObjectId,
+    periodId: Types.ObjectId,
+  ): Promise<PeriodSetting> {
+    const periodSetting = await this.periodSettingsModel
+      .findOne({ period: periodId, setting: settingId })
+      .lean()
+      .populate('period')
+      .populate('setting')
+      .exec();
+
+    if (!periodSetting)
+      throw new ApiException(errorMessages.PERIOD_SETTING_NOT_FOUND);
+    return new PeriodSetting(periodSetting);
+  }
+
+  async setOnePeriodSetting(
+    settingId: Types.ObjectId,
+    periodId: Types.ObjectId,
+    data: SetPeriodSettingDto,
+  ): Promise<PeriodSetting> {
+    const setting = await this.findOneById(settingId);
+    if (!setting) throw new ApiException(errorMessages.SETTING_NOT_FOUND);
+    const { valid, value: validatedValue } = validateSetting(
+      data.value,
+      setting.type,
+    );
+    if (!valid) {
+      throw new ApiException(
+        errorMessages.INVALID_SETTING_VALUE,
+        `Settings value ${data.value} is not valid for type ${setting.type}.`,
+      );
+    }
+
+    const period = await this.periodModel.findById(periodId).lean();
+    if (!period) throw new ApiException(errorMessages.PERIOD_NOT_FOUND);
+    if (period.status !== PeriodStatusType.OPEN)
+      throw new ApiException(
+        errorMessages.PERIOD_SETTING_CAN_BE_CHANGED__WHEN_ITS_OPEN,
+      );
+
+    const periodSetting = await this.periodSettingsModel.findOne({
+      setting: settingId,
+      period: period._id,
+    });
+    if (!periodSetting)
+      throw new ApiException(errorMessages.PERIOD_SETTING_NOT_FOUND);
+
+    const originalValue = periodSetting.value;
+
+    if (typeof data.value === 'undefined') {
+      throw new ApiException(errorMessages.VALUE_IS_REQUIRED_FIELD);
+    }
+    periodSetting.value = validatedValue;
+
+    await periodSetting.save();
+
+    await this.eventLogService.logEvent({
+      typeKey: EventLogTypeKey.SETTING,
+      description: `Updated period "${period.name}" setting "${
+        setting.label
+      }" from "${originalValue || ''}" to "${setting.value || ''}"`,
+    });
+
+    return this.findOneBySettingIdAndPeriodId(settingId, periodId);
+  }
+
+  /**
+   * Create period settings for a period based on the default settings found
+   * in the settings collection, marked with the PERIOD_DEFAULT group.
+   */
+  async createSettingsForPeriod(periodId: Types.ObjectId) {
+    const settingsAllreadyExist = await this.findAllPeriodSettings(periodId);
+    if (settingsAllreadyExist.length > 0) {
+      throw new ApiException(
+        errorMessages.PERIOD_SETTINGS_ALREADY_EXIST_FOR_THIS_PERIOD,
+      );
+    }
+
+    const periodSettingsDefaults = await this.findByGroup(
+      SettingGroup.PERIOD_DEFAULT,
+    );
+
+    const periodSettings = periodSettingsDefaults.map((setting) => {
+      return {
+        period: periodId,
+        setting: setting._id,
+        value: setting.value,
+      };
+    });
+
+    return this.periodSettingsModel.insertMany(periodSettings);
   }
 }
